heating {mlogitBMA}	R Documentation

Heating Dataset

Description

Kenneth Trains dataset containing data on choice of heating system in California houses.

Usage

data(heating)

Format

A data frame with 900 observations on the following 19 variables.

idcase

Observation number.

depvar

Identifies the chosen alternative (1-5).

ic1

Installation cost for a gas central system.

ic2

Installation cost for a gas room system.

ic3

Installation cost for a electric central system.

ic4

Installation cost for a electric room system.

ic5

Installation cost for a heat pump.

oc1

Annual operating cost for a gas central system.

oc2

Annual operating cost for a gas room system.

oc3

Annual operating cost for a electric central system.

oc4

Annual operating cost for a electric room system.

oc5

Annual operating cost for a heat pump.

income

Annual income of the household.

agehed

Age of the household head.

rooms

Number of rooms in the house.

ncost1

Identifies whether the house is in the northern coastal region.

scost1

Identifies whether the house is in the southern coastal region.

mountn

Identifies whether the house is in the mountain region.

valley

Identifies whether the house is in the central valley region.

Details

The observations consist of single-family houses in California that were newly built and had central air-conditioning. The choice is among heating systems. Five types of systems are considered to have been possible:

(1) gas central, (2) gas room, (3) electric central, (4) electric room, (5) heat pump.

For these data, the costs were calculated as the amount the system would cost if it were installed in the house, given the characteristics of the house (such as size), the price of gas and electricity in the house location, and the weather conditions in the area (which determine the necessary capacity of the system and the amount it will be run.) These cost are conditional on the house having central air-conditioning. (That is why the installation cost of gas central is lower than that for gas room: the central system can use the air-conditioning products that have been installed.)

Note

This help file was created using Kenneth Trains description of the dataset, see Source.

Source

http://elsa.berkeley.edu/~train/distant.html

References

Train, K.E. (2003) Discrete Choice Methods with Simulation. Cambridge University Press.

Examples

data(heating)
head(heating)

[Package mlogitBMA version 0.1-9 Index]