mlogitBMA-package {mlogitBMA}	R Documentation

Bayesian Model Averaging for Multinomial Logit Models

Description

Provides a modified function bic.glm of the BMA package that can be applied to multinomial logit (MNL) data. The data is converted to binary logit using the Begg & Gray approximation. The package also contains functions for maximum likelihood estimation of MNL models.

Details

The main function of the package is bic.mlogit which runs the Bayesian Model Averaging on multinomial logit data. Results can be explored using summary.bic.mlogit, imageplot.mlogit, or plot.bic.mlogit functions.

An MNL estimation of a single model can be done using estimate.mlogit. Use summary.mnl to view its results.

Author(s)

Hana Sevcikova, Adrian Raftery

Maintainer: Hana Sevcikova <hanas@uw.edu>

References

Begg, C.B., Gray, R. (1984) Calculation of polychotomous logistic regression parameters using individualized regressions. Biometrika 71, 11–18.

Raftery, A.E. (1995) Bayesian model selection in social research (with Discussion). Sociological Methodology 1995 (Peter V. Marsden, ed.), 111–196, Cambridge, Mass.: Blackwells.

Train, K.E. (2003) Discrete Choice Methods with Simulation. Cambridge University Press.

Yeung, K.Y., Bumgarner, R.E., Raftery, A.E. (2005) Bayesian model averaging: development of an improved multi-class, gene selection and classification tool for microarray data. Bioinformatics 21 (10), 2394–2402.

See Also

bic.glm

[Package mlogitBMA version 0.1-9 Index]