summary.bic.mlogit {mlogitBMA}	R Documentation

Summary and Plotting Functions

Description

Summarizes and plots results of the bic.mlogit function.

Usage

## S3 method for class 'bic.mlogit'
summary(object, ...)

## S3 method for class 'bic.mlogit'
plot(x, ...)

imageplot.mlogit (x , ...)

Arguments

object, x
Object of class bic.mlogit.

...
Arguments passed to the underlying functions.

Details

summary prints a summary of object, using the BMA function summary.bic.glm. It also prints a summary of the model specification, using summary.mnl.spec.

plot produces a plot of the posterior distribution of the coefficients produced by model averaging. It uses the BMA function plot.bic.glm.

imageplot.mlogit creates an image of the selected models, using the BMA function imageplot.bma.

Author(s)

Hana Sevcikova

See Also

bic.mlogit

Examples

# See example in bic.mlogit 

[Package mlogitBMA version 0.1-9 Index]