summary.mnl {mlogitBMA}	R Documentation

Summary for Results of a Multinomial Logit Estimation

Description

Gives a summary for an object of class mnl which contains results of a multinomial logit estimation.

Usage

## S3 method for class 'mnl'
 summary(object, ...)

Arguments

object
Object of class mnl

...
Not used.

Author(s)

Hana Sevcikova

[Package mlogitBMA version 0.1-9 Index]